requirements.txt
wand
pillow
fpdf